fix: address explorer review feedback
