+++ doc/.governance/rust-oversized-file-baseline.tsv
@@ -7,7 +7,6 @@ code	crates/oasis7_node/src/replication.rs	1221
 code	crates/oasis7_node/src/types.rs	1277
 test	crates/oasis7/src/simulator/llm_agent/tests_split_part1.rs	1204
 test	crates/oasis7/src/viewer/runtime_live/tests/auth_actions.rs	1335
-test	crates/oasis7_client_launcher/src/main_tests.rs	1225
 test	crates/oasis7_net/src/tests.rs	1239
 test	crates/oasis7_node/src/tests_split_part1.rs	1427
 test	crates/oasis7_node/src/tests_split_part3.rs	1735
